// src/types/MarketInfo.ts
export type MarketInfo = {
  name: string;
  phone: string;
  marketName: string;
  postalAddress: string;
  defaultAddress: string;
  restAddress: string;
  registNum: string;
  RegistFile: File | null;
  Passbook: File | null;
  certifidoc: File | null;
  bankNum: string;
  bank: string;
  intro: string;
  profile: File | null;

};